requirements.txt
langchain
openai
huggingface_hub
python-dotenv
PyPDF2
openai
langchain_community
cassio
datasets

